requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.1.0
python-multipart==0.0.20
httpx==0.28.1
langgraph==0.3.29